Drew - 真。真。刪除商品

+++ src/main/java/foodelicious/compbackendrest/controller/CompanyBackEndRestController.java
@@ -67,7 +67,7 @@ public class CompanyBackEndRestController {
 	}
 
 	@DeleteMapping("/companyProducts/delete/{productId}")
-	public String deleteProduct(@PathVariable Long productId) {
+	public String deleteProduct(@PathVariable Integer productId) {
 
 		return cbkProductDao.deleteProduct(productId);
 	}
